# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "HS-SPORT API",
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/player": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Get player",
                "description": "Get player",
                "operationId": "api.player",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "sizeSystem",
                        "in": "query",
                        "description": "size system name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/player/about": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "About",
                "description": "About",
                "operationId": "api.player.about",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "sizeSystem",
                        "in": "query",
                        "description": "size system name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/player/about-stats": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "About - stats",
                "description": "About - stats",
                "operationId": "api.player.about.stats",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/player/player-standings": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Player Standings",
                "description": "Player Standings",
                "operationId": "api.player.player.standings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "viewType",
                        "in": "query",
                        "description": "state || school || team",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/player/player-standings-page": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Get Player Standings - pagination number",
                "description": "Get Player Standings - pagination number",
                "operationId": "api.player.player.standings.page",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "viewType",
                        "in": "query",
                        "description": "state || school || team",
                        "required": true,
                        "type": "string",
                        "default": "state"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": true,
                        "type": "integer",
                        "default": 2
                    },
                    {
                        "name": "isFavorite",
                        "in": "query",
                        "description": "use favorite filter",
                        "required": true,
                        "type": "boolean",
                        "default": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/player/player-standings-search": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Player Standings - Search",
                "description": "Player Standings - Search",
                "operationId": "api.player.player.standings.search",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "viewType",
                        "in": "query",
                        "description": "state || school || team",
                        "required": true,
                        "type": "string",
                        "default": "state"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "query search",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "isFavorite",
                        "in": "query",
                        "description": "use favorite filter",
                        "required": true,
                        "type": "boolean",
                        "default": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/statistics/standings": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Team Standings",
                "description": "Stats - Team Standings",
                "operationId": "api.statistics.standings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "1 or 2",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "1 or 2 or null",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "team ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "stateId",
                        "in": "query",
                        "description": "state ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "city ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}